requirements.txt
beautifulsoup4==4.9.1
Django==2.2.16
django-autocomplete-light==3.4.1
django-bootstrap-static==4.2.1
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-markdownx==3.0.1
django-ordered-model==3.4.1
django-responsive-images==1.2.0
django-static-fontawesome==5.10.2
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==6.1.0
pymdown-extensions==8.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==1.2.0
